# docs/swagger.yaml
basePath: /
definitions:
...
  merchant.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  models.CouponTemplate:
    properties:
      createdAt:
        description: IsActive      bool    `gorm:"default:true;not null"`
        type: string
      description:
        type: string
      discountType:
        description: fixed, percent
        type: string
      discountValue:
        type: integer
      id:
        type: integer
      merchantID:
        type: integer
      minAmount:
        type: integer
      name:
        type: string
      totalCount:
        type: integer
      updatedAt:
        type: string
      validityDays:
        type: integer
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 预约系统API文档
  title: Appointment System API
  version: "1.0"
paths:
...
  /api/merchant/coupons:
    get:
      description: 获取当前商户的所有优惠券模板（商户端）
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页码
        example: 1
        in: query
        name: page
        type: integer
      - description: 每页数量
        example: 20
        in: query
        name: page_size
        type: integer
      - description: 状态
        enum:
        - active
        - expired
        example: active
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回优惠券模板列表
          schema:
            items:
              $ref: '#/definitions/models.CouponTemplate'
            type: array
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取优惠券模板列表
      tags:
      - 商户-优惠券管理
    post:
      consumes:
      - application/json
      description: 创建新的优惠券模板（商户端）
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 优惠券模板信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CouponTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回创建的优惠券模板
          schema:
            $ref: '#/definitions/models.CouponTemplate'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建优惠券模板
      tags:
      - 商户-优惠券管理
  /api/merchant/coupons/{couponTemplateId}:
    delete:
      consumes:
      - application/json
      description: 删除指定的优惠券模板（商户端）。注意：只能删除未被使用的优惠券模板。
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 优惠券模板ID
        example: 123
        in: path
        name: couponTemplateId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 无效的优惠券ID或优惠券已被使用
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 优惠券不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除优惠券模板
      tags:
      - 商户-优惠券管理
    put:
      consumes:
      - application/json
      description: 更新指定的优惠券模板（商户端）
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 优惠券模板ID
        example: 123
        in: path
        name: couponTemplateId
        required: true
        type: integer
      - description: 优惠券模板更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CouponTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 优惠券不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新优惠券模板
      tags:
      - 商户-优惠券管理
  /api/merchant/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录，获取JWT令牌
      parameters:
      - description: 登录凭证
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/merchant.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '登录成功" Example({"token": "jwt_token", "admin": {"id":"1", "username":"admin",
            "role":"super_admin", "merchant_id":"123"}})'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: '参数错误" Example({"error": "参数错误"})'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: '用户名或密码错误" Example({"error": "用户名或密码错误"})'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: '生成token失败" Example({"error": "生成token失败"})'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 商户管理员登录
      tags:
      - 认证
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"